splitted runtime evaluator in separate class
